Abstract error components into separate files
